requirements.txt
aiohttp
torch
torchvision
#opencv-python
ultralytics
scikit-image
asgiref==3.8.1
Django==4.1.13
djongo==1.3.6
dnspython==2.7.0
pillow==11.0.0
pymongo==4.10.1
sqlparse==0.2.4
numpy==2.1.3
opencv-python-headless==4.10.0.84
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
onnxruntime
pyyaml